requirements.txt
python-telegram-bot==22.5
faster-whisper
sumy
nltk
